requirements/development.txt
# Development requirements

-r base.txt

django-debug-toolbar==2.2